05 Jul 23:43
I was kind of locked in this time so i forgot to commit until now lol

- added some neat effects and animations to the "locate" and "clear" buttons
- added a moving gradient bg to the "main-content" section
- added a favicon
- added copy functionality to all leylines
- made the header sticky
- made focus shift to the next input when user presses enter, then call the calculate function
- modified the "--success" colour a little bit
- slight adjustments to the info (i might make it look better at some point idk)
- some slight code adjustments

i might do a big round of code optimization sometime, but that won't be for a while

also the mobile mode is... ok for now. i might improve it later. kind of sad that the cool animations don't work super well on mobile :\